fix(phase-1): fix CI build errors and devcontainer environment
- RenderSystem.cpp: remove params.Stencil (not a member of
  SIrrlichtCreationParameters in this Irrlicht vcpkg port)
- IrrlichtRenderer.cpp: cast getAnimators().size() to size_t to fix
  %zu format mismatch warning (-Wformat treated as error)
- ui_smoke_test.cpp: fix StubUIBackend to match current IUIBackend
  interface (std::string not wstring, correct param order for addButton,
  correct setElementImage signature, correct getElementText return type,
  add missing getVirtualWidth/getVirtualHeight/loadTexture overrides)
- ci.yml: move Windows GLEW verification steps to after cmake configure
  (build/vcpkg_installed/ is created during cmake configure via vcpkg
  manifest mode — checking before configure always fails)
- Dockerfile: add libglu1-mesa-dev (required for vcpkg glew pkgconfig
  fixup) and gcc-13/g++-13 (required for openal-soft 1.25.0 <format>)

+++ src/rendering/IrrlichtRenderer.cpp
@@ -74,7 +74,7 @@ void IrrlichtRenderer::setCamera(const CameraParams& p) {
                 fprintf(stderr,
                     "[IrrlichtRenderer] WARNING: unexpected animators on addCameraSceneNode() "
                     "result — removing %zu animator(s)\n",
-                    m_camera->getAnimators().size());
+                    static_cast<size_t>(m_camera->getAnimators().size()));
             }
 #endif
             while (m_camera->getAnimators().size() > 0) {
